Design Left Side of Signup Page
I am only using Chrome for now since other browsers render it differently. Absolute and relative positions were applied for the credits and for the banner. I also learned that you do not have to only use flex. The display in-line block was very effective in ensuring that the logo and the text are on the same axis. I also downloaded and used a custom font for the brand text. The CCS of the credits are very clean.

TODO:
- fix flex grow and shrink of the 2 sides
- stylize the right side of the page
- make it consistent across Firefox and Chrome